requirements.txt
playwright
google-generativeai
pydantic
pandas
python-dotenv
importlib-metadata
